\name{cadswes-package}
\alias{cadswes-package}
\alias{cadswes}
\docType{package}

\title{
Utilities developed at the Center for Advanced Decision Support for Water and Environmental Systems (CADSWES)
}

\description{
Functions for various purposes including water resources data analysis and interacting with Riverware data.
}

\details{
\tabular{ll}{
Package: \tab cadswes\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-05-09\cr
License: \tab GPL 3.0 or higher\cr
LazyLoad: \tab yes\cr
}
% ~~ An overview of how to use the package, including the most important ~~
% ~~ functions ~~
}

\author{
Cameron Bracken <cameron.bracken@colorado.edu>
}


\keyword{ package }
\seealso{
\code{\link[cadswes]{read.rdf}}\cadswes{\link[cadswes]{read_rdf_2}}
}
\examples{
# 
}
