\name{read.rdf}
\alias{read.rdf}
\title{Read a Riverware RDF file and save it to a list...}
\usage{read.rdf(file)}

\description{
  Read a Riverware RDF file and save it to a list
}
\value{a list containing the contents of the RDF file.}
\note{Due to the structure of RDF files, large RDF files may take a LONG
time to read in.}
\author{Cameron Bracken <cameron.bracken@gmail.com>}
\arguments{
  \item{file}{An RDF file}
}
\examples{# None for now}
